+++ server/src/main/java/com/linbit/linstor/core/apis/ResourceWithPayloadApi.java
@@ -15,4 +15,7 @@ public interface ResourceWithPayloadApi
 
     @Nullable
     Integer getPortCount();
+
+    @Nullable
+    Boolean isDrbdClient();
 }